[Decomp][Chat] Cohesion split: extract arg-extraction/validation/messaging/help to ChatXxx.cpp (#252)
Re-applies the mangosthree decomposition onto MangosOne's own TBC 2.4.3
source. Pure code move out of the parent translation unit into the new
cohesively-named file(s); no behaviour change (source diff is deletion-only).
CMake file(GLOB) picks the new TU(s) up automatically. Verified: game.lib
compiles and links clean (MSVC, Release).

Not ported (absent in MangosOne TBC 2.4.3 -- these are
later-expansion methods that do not exist in this branch):
  ChatHandler::ExtractUInt64,ChatHandler::ExtractInt64

ChatHandler::ExtractUInt64,ChatHandler::ExtractInt64
Ported-from: mangosthree/server@26063e2606bb
