+++ src/game/WorldHandlers/Chat.cpp
@@ -1079,152 +1079,6 @@ bool ChatHandler::hasStringAbbr(const char* name, const char* part)
 }
 
 /**
- * @brief Sends a system message to the current session, splitting multiline text.
- *
- * @param str The message text to send.
- */
-void ChatHandler::SendSysMessage(const char* str)
-{
-    WorldPacket data;
-
-    // need copy to prevent corruption by strtok call in LineFromMessage original string
-    char* buf = mangos_strdup(str);
-    char* pos = buf;
-
-    while (char* line = LineFromMessage(pos))
-    {
-        // m_session == null when we're accessing these command from the console.
-        ObjectGuid senderGuid;
-        if (m_session)
-        {
-            senderGuid = m_session->GetPlayer()->GetObjectGuid();
-        }
-
-        ChatHandler::BuildChatPacket(data, CHAT_MSG_SYSTEM, line, LANG_UNIVERSAL, CHAT_TAG_NONE, senderGuid);
-        m_session->SendPacket(&data);
-    }
-
-    delete[] buf;
-}
-
-/**
- * @brief Sends a global system message to all sessions meeting the minimum security level.
- *
- * @param str The message text to send.
- * @param minSec The minimum security level required to receive the message.
- */
-void ChatHandler::SendGlobalSysMessage(const char* str, AccountTypes minSec)
-{
-    // Chat output
-    WorldPacket data;
-
-    // need copy to prevent corruption by strtok call in LineFromMessage original string
-    char* buf = mangos_strdup(str);
-    char* pos = buf;
-    ObjectGuid senderGuid = m_session ? m_session->GetPlayer()->GetObjectGuid() : ObjectGuid();
-
-    while (char* line = LineFromMessage(pos))
-    {
-        // m_session == null when we're accessing these command from the console.
-        ObjectGuid senderGuid;
-        if (m_session)
-        {
-            senderGuid = m_session->GetPlayer()->GetObjectGuid();
-        }
-
-        ChatHandler::BuildChatPacket(data, CHAT_MSG_SYSTEM, line, LANG_UNIVERSAL, CHAT_TAG_NONE, senderGuid);
-        sWorld.SendGlobalMessage(&data, minSec);
-    }
-
-    delete[] buf;
-}
-
-/**
- * @brief Sends a localized system message by string table entry.
- *
- * @param entry The string table entry identifier.
- */
-void ChatHandler::SendSysMessage(int32 entry)
-{
-    SendSysMessage(GetMangosString(entry));
-}
-
-/**
- * @brief Formats and sends a localized system message.
- *
- * @param entry The string table entry identifier.
- */
-void ChatHandler::PSendSysMessage(int32 entry, ...)
-{
-    const char* format = GetMangosString(entry);
-    va_list ap;
-    char str [2048];
-    va_start(ap, entry);
-    vsnprintf(str, 2048, format, ap);
-    va_end(ap);
-    SendSysMessage(str);
-}
-
-/**
- * @brief Formats and sends a localized multiline system message using @@ separators.
- *
- * @param entry The string table entry identifier.
- */
-void  ChatHandler::PSendSysMessageMultiline(int32 entry, ...)
-{
-    uint32 linecount = 0;
-
-    const char* format = GetMangosString(entry);
-    va_list ap;
-    char str[2048];
-    va_start(ap, entry);
-    vsnprintf(str, 2048, format, ap);
-    va_end(ap);
-
-    std::string mangosString(str);
-
-    /* Used for tracking our position within the string while iterating through it */
-    std::string::size_type pos = 0, nextpos;
-
-    /* Find the next occurance of @ in the string
-     * This is how newlines are represented */
-    while ((nextpos = mangosString.find("@@", pos)) != std::string::npos)
-    {
-        /* If these are not equal, it means a '@@' was found
-         * These are used to represent newlines in the string
-         * It is set by the code above here */
-        if (nextpos != pos)
-        {
-            /* Send the player a system message containing the substring from pos to nextpos - pos */
-            PSendSysMessage("%s", mangosString.substr(pos, nextpos - pos).c_str());
-            ++linecount;
-        }
-        pos = nextpos + 2; // +2 because there are two @ as delimiter
-    }
-
-    /* There are no more newlines in our mangosString, so we send whatever is left */
-    if (pos < mangosString.length())
-    {
-        PSendSysMessage("%s", mangosString.substr(pos).c_str());
-    }
-}
-
-/**
- * @brief Formats and sends a raw system message.
- *
- * @param format The printf-style format string.
- */
-void ChatHandler::PSendSysMessage(const char* format, ...)
-{
-    va_list ap;
-    char str [2048];
-    va_start(ap, format);
-    vsnprintf(str, 2048, format, ap);
-    va_end(ap);
-    SendSysMessage(str);
-}
-
-/**
  * @brief Validates command table structure and reports inconsistencies.
  *
  * @param table The command table to validate.
@@ -1652,708 +1506,6 @@ bool ChatHandler::ParseCommands(const char* text)
 }
 
 /**
- * @brief Displays the available subcommands for a command table.
- *
- * @param table The subcommand table to display.
- * @param cmd The parent command name.
- * @return true if any subcommands were shown; otherwise false.
- */
-bool ChatHandler::ShowHelpForSubCommands(ChatCommand* table, char const* cmd)
-{
-    std::string list;
-    for (uint32 i = 0; table[i].Name != NULL; ++i)
-    {
-        // must be available (ignore handler existence for show command with possible available subcommands
-        if (!isAvailable(table[i]))
-        {
-            continue;
-        }
-
-        if (m_session)
-        {
-            list += "\n    ";
-        }
-        else
-        {
-            list += "\n\r    ";
-        }
-
-        list += table[i].Name;
-
-        if (table[i].ChildCommands)
-        {
-            list += " ...";
-        }
-    }
-
-    if (list.empty())
-    {
-        return false;
-    }
-
-    if (table == getCommandTable())
-    {
-        SendSysMessage(LANG_AVIABLE_CMD);
-        SendSysMessage(list.c_str());
-    }
-    else
-    {
-        PSendSysMessage(LANG_SUBCMDS_LIST, cmd);
-        SendSysMessage(list.c_str());
-    }
-    return true;
-}
-
-/**
- * @brief Displays help text and subcommands for a command path.
- *
- * @param table The command table to search.
- * @param cmd The command path to resolve.
- * @return true if help content was shown; otherwise false.
- */
-bool ChatHandler::ShowHelpForCommand(ChatCommand* table, const char* cmd)
-{
-    char const* oldCmd = cmd;
-    ChatCommand* command = NULL;
-    ChatCommand* parentCommand = NULL;
-
-    ChatCommand* showCommand = NULL;
-    ChatCommand* childCommands = NULL;
-
-    ChatCommandSearchResult res = FindCommand(table, cmd, command, &parentCommand);
-
-    switch (res)
-    {
-        case CHAT_COMMAND_OK:
-        {
-            // for "" subcommand use parent command if any for subcommands list output
-            if (strlen(command->Name) == 0 && parentCommand)
-            {
-                showCommand = parentCommand;
-                cmd = "";
-            }
-            else
-            {
-                showCommand = command;
-            }
-
-            childCommands = showCommand->ChildCommands;
-            break;
-        }
-        case CHAT_COMMAND_UNKNOWN_SUBCOMMAND:
-            showCommand = command;
-            childCommands = showCommand->ChildCommands;
-            break;
-        case CHAT_COMMAND_UNKNOWN:
-            // not show command list at error in first level command find fail
-            childCommands = table != getCommandTable() || strlen(oldCmd) == 0 ? table : NULL;
-            command = NULL;
-            break;
-    }
-
-    if (command && !command->Help.empty())
-    {
-        std::string helpText = command->Help;
-
-        // Attemp to localize help text if not in CLI mode
-        if (m_session)
-        {
-            int loc_idx = m_session->GetSessionDbLocaleIndex();
-            sCommandMgr.GetCommandHelpLocaleString(command->Id, loc_idx, &helpText);
-        }
-
-        SendSysMessage(helpText.c_str());
-    }
-
-    if (childCommands)
-        if (ShowHelpForSubCommands(childCommands, showCommand ? showCommand->Name : ""))
-        {
-            return true;
-        }
-
-    if (command && command->Help.empty())
-    {
-        SendSysMessage(LANG_NO_HELP_CMD);
-    }
-
-    return command || childCommands;
-}
-
-/**
- * @brief Validates chat text and embedded shift-links against configured strictness rules.
- *
- * @param message The message text to validate.
- * @return true if the message is valid; otherwise false.
- */
-bool ChatHandler::isValidChatMessage(const char* message)
-{
-    /*
-
-    valid examples:
-    |cffa335ee|Hitem:812:0:0:0:0:0:0:0:70|h[Glowing Brightwood Staff]|h|r
-    |cff808080|Hquest:2278:47|h[The Platinum Discs]|h|r
-    |cff4e96f7|Htalent:2232:-1|h[Taste for Blood]|h|r
-    |cff71d5ff|Hspell:21563|h[Command]|h|r
-    |cffffd000|Henchant:3919|h[Engineering: Rough Dynamite]|h|r
-
-    | will be escaped to ||
-    */
-
-    if (strlen(message) > 255)
-    {
-        return false;
-    }
-
-    const char validSequence[6] = "cHhhr";
-    const char* validSequenceIterator = validSequence;
-
-    // more simple checks
-    if (sWorld.getConfig(CONFIG_UINT32_CHAT_STRICT_LINK_CHECKING_SEVERITY) < 3)
-    {
-        const std::string validCommands = "cHhr|";
-
-        while (*message)
-        {
-            // find next pipe command
-            message = strchr(message, '|');
-
-            if (!message)
-            {
-                return true;
-            }
-
-            ++message;
-            char commandChar = *message;
-            if (validCommands.find(commandChar) == std::string::npos)
-            {
-                return false;
-            }
-
-            ++message;
-            // validate sequence
-            if (sWorld.getConfig(CONFIG_UINT32_CHAT_STRICT_LINK_CHECKING_SEVERITY) == 2)
-            {
-                if (commandChar == *validSequenceIterator)
-                {
-                    if (validSequenceIterator == validSequence + 4)
-                    {
-                        validSequenceIterator = validSequence;
-                    }
-                    else
-                    {
-                        ++validSequenceIterator;
-                    }
-                }
-                else if (commandChar != '|')
-                {
-                    return false;
-                }
-            }
-        }
-        return true;
-    }
-
-    std::istringstream reader(message);
-    char buffer[256];
-
-    uint32 color = 0;
-
-    ItemPrototype const* linkedItem = NULL;
-    Quest const* linkedQuest = NULL;
-    SpellEntry const* linkedSpell = NULL;
-    ItemRandomPropertiesEntry const* itemProperty = NULL;
-    ItemRandomSuffixEntry const* itemSuffix = NULL;
-
-    while (!reader.eof())
-    {
-        if (validSequence == validSequenceIterator)
-        {
-            linkedItem = NULL;
-            linkedQuest = NULL;
-            linkedSpell = NULL;
-            itemProperty = NULL;
-            itemSuffix = NULL;
-
-            reader.ignore(255, '|');
-        }
-        else if (reader.get() != '|')
-        {
-            DEBUG_LOG("ChatHandler::isValidChatMessage sequence aborted unexpectedly");
-            return false;
-        }
-
-        // pipe has always to be followed by at least one char
-        if (reader.peek() == '\0')
-        {
-            DEBUG_LOG("ChatHandler::isValidChatMessage pipe followed by \\0");
-            return false;
-        }
-
-        // no further pipe commands
-        if (reader.eof())
-        {
-            break;
-        }
-
-        char commandChar;
-        reader >> commandChar;
-
-        // | in normal messages is escaped by ||
-        if (commandChar != '|')
-        {
-            if (commandChar == *validSequenceIterator)
-            {
-                if (validSequenceIterator == validSequence + 4)
-                {
-                    validSequenceIterator = validSequence;
-                }
-                else
-                {
-                    ++validSequenceIterator;
-                }
-            }
-            else
-            {
-                DEBUG_LOG("ChatHandler::isValidChatMessage invalid sequence, expected %c but got %c", *validSequenceIterator, commandChar);
-                return false;
-            }
-        }
-        else if (validSequence != validSequenceIterator)
-        {
-            // no escaped pipes in sequences
-            DEBUG_LOG("ChatHandler::isValidChatMessage got escaped pipe in sequence");
-            return false;
-        }
-
-        switch (commandChar)
-        {
-            case 'c':
-                color = 0;
-                // validate color, expect 8 hex chars
-                for (int i = 0; i < 8; ++i)
-                {
-                    char c;
-                    reader >> c;
-                    if (!c)
-                    {
-                        DEBUG_LOG("ChatHandler::isValidChatMessage got \\0 while reading color in |c command");
-                        return false;
-                    }
-
-                    color <<= 4;
-                    // check for hex char
-                    if (c >= '0' && c <= '9')
-                    {
-                        color |= c - '0';
-                        continue;
-                    }
-                    if (c >= 'a' && c <= 'f')
-                    {
-                        color |= 10 + c - 'a';
-                        continue;
-                    }
-                    DEBUG_LOG("ChatHandler::isValidChatMessage got non hex char '%c' while reading color", c);
-                    return false;
-                }
-                break;
-            case 'H':
-                // read chars up to colon  = link type
-                reader.getline(buffer, 256, ':');
-                if (reader.eof())                           // : must be
-                {
-                    return false;
-                }
-
-                if (strcmp(buffer, "item") == 0)
-                {
-                    // read item entry
-                    reader.getline(buffer, 256, ':');
-                    if (reader.eof())                       // : must be
-                    {
-                        return false;
-                    }
-
-                    linkedItem = ObjectMgr::GetItemPrototype(atoi(buffer));
-                    if (!linkedItem)
-                    {
-                        DEBUG_LOG("ChatHandler::isValidChatMessage got invalid itemID %u in |item command", atoi(buffer));
-                        return false;
-                    }
-
-                    if (color != ItemQualityColors[linkedItem->Quality])
-                    {
-                        DEBUG_LOG("ChatHandler::isValidChatMessage linked item has color %u, but user claims %u", ItemQualityColors[linkedItem->Quality],
-                                  color);
-                        return false;
-                    }
-
-                    // the itementry is followed by several integers which describe an instance of this item
-
-                    // position relative after itemEntry
-                    const uint8 randomPropertyPosition = 6;
-
-                    int32 propertyId = 0;
-                    bool negativeNumber = false;
-                    char c;
-                    for (uint8 i = 0; i < randomPropertyPosition; ++i)
-                    {
-                        propertyId = 0;
-                        negativeNumber = false;
-                        while ((c = reader.get()) != ':')
-                        {
-                            if (c >= '0' && c <= '9')
-                            {
-                                propertyId *= 10;
-                                propertyId += c - '0';
-                            }
-                            else if (c == '-')
-                            {
-                                negativeNumber = true;
-                            }
-                            else
-                            {
-                                return false;
-                            }
-                        }
-                    }
-                    if (negativeNumber)
-                    {
-                        propertyId *= -1;
-                    }
-
-                    if (propertyId > 0)
-                    {
-                        itemProperty = sItemRandomPropertiesStore.LookupEntry(propertyId);
-                        if (!itemProperty)
-                        {
-                            return false;
-                        }
-                    }
-                    else if (propertyId < 0)
-                    {
-                        itemSuffix = sItemRandomSuffixStore.LookupEntry(-propertyId);
-                        if (!itemSuffix)
-                        {
-                            return false;
-                        }
-                    }
-
-                    // ignore other integers
-                    while ((c >= '0' && c <= '9') || c == ':')
-                    {
-                        reader.ignore(1);
-                        c = reader.peek();
-                    }
-                }
-                else if (strcmp(buffer, "quest") == 0)
-                {
-                    // no color check for questlinks, each client will adapt it anyway
-                    uint32 questid = 0;
-                    // read questid
-                    char c = reader.peek();
-                    while (c >= '0' && c <= '9')
-                    {
-                        reader.ignore(1);
-                        questid *= 10;
-                        questid += c - '0';
-                        c = reader.peek();
-                    }
-
-                    linkedQuest = sObjectMgr.GetQuestTemplate(questid);
-
-                    if (!linkedQuest)
-                    {
-                        DEBUG_LOG("ChatHandler::isValidChatMessage Questtemplate %u not found", questid);
-                        return false;
-                    }
-
-                    if (c != ':')
-                    {
-                        DEBUG_LOG("ChatHandler::isValidChatMessage Invalid quest link structure");
-                        return false;
-                    }
-
-                    reader.ignore(1);
-                    c = reader.peek();
-                    // level
-                    uint32 questlevel = 0;
-                    while (c >= '0' && c <= '9')
-                    {
-                        reader.ignore(1);
-                        questlevel *= 10;
-                        questlevel += c - '0';
-                        c = reader.peek();
-                    }
-
-                    if (questlevel >= STRONG_MAX_LEVEL)
-                    {
-                        DEBUG_LOG("ChatHandler::isValidChatMessage Quest level %u too big", questlevel);
-                        return false;
-                    }
-
-                    if (c != '|')
-                    {
-                        DEBUG_LOG("ChatHandler::isValidChatMessage Invalid quest link structure");
-                        return false;
-                    }
-                }
-                else if (strcmp(buffer, "talent") == 0)
-                {
-                    // talent links are always supposed to be blue
-                    if (color != CHAT_LINK_COLOR_TALENT)
-                    {
-                        return false;
-                    }
-
-                    // read talent entry
-                    reader.getline(buffer, 256, ':');
-                    if (reader.eof())                       // : must be
-                    {
-                        return false;
-                    }
-
-                    TalentEntry const* talentInfo = sTalentStore.LookupEntry(atoi(buffer));
-                    if (!talentInfo)
-                    {
-                        return false;
-                    }
-
-                    linkedSpell = sSpellStore.LookupEntry(talentInfo->RankID[0]);
-                    if (!linkedSpell)
-                    {
-                        return false;
-                    }
-
-                    char c = reader.peek();
-                    // skillpoints? whatever, drop it
-                    while (c != '|' && c != '\0')
-                    {
-                        reader.ignore(1);
-                        c = reader.peek();
-                    }
-                }
-                else if (strcmp(buffer, "spell") == 0)
-                {
-                    if (color != CHAT_LINK_COLOR_SPELL)
-                    {
-                        return false;
-                    }
-
-                    uint32 spellid = 0;
-                    // read spell entry
-                    char c = reader.peek();
-                    while (c >= '0' && c <= '9')
-                    {
-                        reader.ignore(1);
-                        spellid *= 10;
-                        spellid += c - '0';
-                        c = reader.peek();
-                    }
-                    linkedSpell = sSpellStore.LookupEntry(spellid);
-                    if (!linkedSpell)
-                    {
-                        return false;
-                    }
-                }
-                else if (strcmp(buffer, "enchant") == 0)
-                {
-                    if (color != CHAT_LINK_COLOR_ENCHANT)
-                    {
-                        return false;
-                    }
-
-                    uint32 spellid = 0;
-                    // read spell entry
-                    char c = reader.peek();
-                    while (c >= '0' && c <= '9')
-                    {
-                        reader.ignore(1);
-                        spellid *= 10;
-                        spellid += c - '0';
-                        c = reader.peek();
-                    }
-                    linkedSpell = sSpellStore.LookupEntry(spellid);
-                    if (!linkedSpell)
-                    {
-                        return false;
-                    }
-                }
-                else
-                {
-                    DEBUG_LOG("ChatHandler::isValidChatMessage user sent unsupported link type '%s'", buffer);
-                    return false;
-                }
-                break;
-            case 'h':
-                // if h is next element in sequence, this one must contain the linked text :)
-                if (*validSequenceIterator == 'h')
-                {
-                    // links start with '['
-                    if (reader.get() != '[')
-                    {
-                        DEBUG_LOG("ChatHandler::isValidChatMessage link caption doesn't start with '['");
-                        return false;
-                    }
-                    reader.getline(buffer, 256, ']');
-                    if (reader.eof())                       // ] must be
-                    {
-                        return false;
-                    }
-
-                    // verify the link name
-                    if (linkedSpell)
-                    {
-                        // spells with that flag have a prefix of "$PROFESSION: "
-                        if (linkedSpell->HasAttribute(SPELL_ATTR_TRADESPELL))
-                        {
-                            // lookup skillid
-                            SkillLineAbilityMapBounds bounds = sSpellMgr.GetSkillLineAbilityMapBounds(linkedSpell->Id);
-                            if (bounds.first == bounds.second)
-                            {
-                                return false;
-                            }
-
-                            SkillLineAbilityEntry const* skillInfo = bounds.first->second;
-
-                            if (!skillInfo)
-                            {
-                                return false;
-                            }
-
-                            SkillLineEntry const* skillLine = sSkillLineStore.LookupEntry(skillInfo->skillId);
-                            if (!skillLine)
-                            {
-                                return false;
-                            }
-
-                            for (uint8 i = 0; i < MAX_LOCALE; ++i)
-                            {
-                                uint32 skillLineNameLength = strlen(skillLine->name[i]);
-                                if (skillLineNameLength > 0 && strncmp(skillLine->name[i], buffer, skillLineNameLength) == 0)
-                                {
-                                    // found the prefix, remove it to perform spellname validation below
-                                    // -2 = strlen(": ")
-                                    uint32 spellNameLength = strlen(buffer) - skillLineNameLength - 2;
-                                    memmove(buffer, buffer + skillLineNameLength + 2, spellNameLength + 1);
-                                }
-                            }
-                        }
-                        bool foundName = false;
-                        for (uint8 i = 0; i < MAX_LOCALE; ++i)
-                        {
-                            if (*linkedSpell->SpellName[i] && strcmp(linkedSpell->SpellName[i], buffer) == 0)
-                            {
-                                foundName = true;
-                                break;
-                            }
-                        }
-                        if (!foundName)
-                        {
-                            return false;
-                        }
-                    }
-                    else if (linkedQuest)
-                    {
-                        if (linkedQuest->GetTitle() != buffer)
-                        {
-                            QuestLocale const* ql = sObjectMgr.GetQuestLocale(linkedQuest->GetQuestId());
-
-                            if (!ql)
-                            {
-                                DEBUG_LOG("ChatHandler::isValidChatMessage default questname didn't match and there is no locale");
-                                return false;
-                            }
-
-                            bool foundName = false;
-                            for (uint8 i = 0; i < ql->Title.size(); ++i)
-                            {
-                                if (ql->Title[i] == buffer)
-                                {
-                                    foundName = true;
-                                    break;
-                                }
-                            }
-                            if (!foundName)
-                            {
-                                DEBUG_LOG("ChatHandler::isValidChatMessage no quest locale title matched");
-                                return false;
-                            }
-                        }
-                    }
-                    else if (linkedItem)
-                    {
-                        char* const* suffix = itemSuffix ? itemSuffix->nameSuffix : (itemProperty ? itemProperty->nameSuffix : NULL);
-
-                        std::string expectedName = std::string(linkedItem->Name1);
-                        if (suffix)
-                        {
-                            expectedName += " ";
-                            expectedName += suffix[LOCALE_enUS];
-                        }
-
-                        if (expectedName != buffer)
-                        {
-                            ItemLocale const* il = sObjectMgr.GetItemLocale(linkedItem->ItemId);
-
-                            bool foundName = false;
-                            for (uint8 i = LOCALE_koKR; i < MAX_LOCALE; ++i)
-                            {
-                                int8 dbIndex = sObjectMgr.GetIndexForLocale(LocaleConstant(i));
-                                if (dbIndex == -1 || il == NULL || (size_t)dbIndex >= il->Name.size())
-                                    // using strange database/client combinations can lead to this case
-                                {
-                                    expectedName = linkedItem->Name1;
-                                }
-                                else
-                                {
-                                    expectedName = il->Name[dbIndex];
-                                }
-                                if (suffix)
-                                {
-                                    expectedName += " ";
-                                    expectedName += suffix[i];
-                                }
-                                if (expectedName == buffer)
-                                {
-                                    foundName = true;
-                                    break;
-                                }
-                            }
-                            if (!foundName)
-                            {
-                                DEBUG_LOG("ChatHandler::isValidChatMessage linked item name wasn't found in any localization");
-                                return false;
-                            }
-                        }
-                    }
-                    // that place should never be reached - if nothing linked has been set in |H
-                    // it will return false before
-                    else
-                    {
-                        return false;
-                    }
-                }
-                break;
-            case 'r':
-            case '|':
-                // no further payload
-                break;
-            default:
-                DEBUG_LOG("ChatHandler::isValidChatMessage got invalid command |%c", commandChar);
-                return false;
-        }
-    }
-
-    // check if every opened sequence was also closed properly
-    if (validSequence != validSequenceIterator)
-    {
-        DEBUG_LOG("ChatHandler::isValidChatMessage EOF in active sequence");
-    }
-
-    return validSequence == validSequenceIterator;
-}
-
-/**
  * @brief Returns the currently selected player, or the session player if nothing is selected.
  *
  * @return Player* The selected player, or NULL when no player can be resolved.
@@ -2423,1647 +1575,6 @@ Creature* ChatHandler::getSelectedCreature()
     }
 
     return m_session->GetPlayer()->GetMap()->GetAnyTypeCreature(m_session->GetPlayer()->GetSelectionGuid());
-}
-
-/**
- * Function skip all whitespaces in args string
- *
- * @param args variable pointer to non parsed args string, updated at function call to new position (with skipped white spaces)
- *             allowed NULL string pointer stored in *args
- */
-void ChatHandler::SkipWhiteSpaces(char** args)
-{
-    if (!*args)
-    {
-        return;
-    }
-
-    while (isWhiteSpace(**args))
-    {
-        ++(*args);
-    }
-}
-
-/**
- * Function extract to val arg signed integer value or fail
- *
- * @param args variable pointer to non parsed args string, updated at function call to new position (with skipped white spaces)
- * @param val  return extracted value if function success, in fail case original value unmodified
- * @return     true if value extraction successful
- */
-bool  ChatHandler::ExtractInt32(char** args, int32& val)
-{
-    if (!*args || !** args)
-    {
-        return false;
-    }
-
-    char* tail = *args;
-
-    long valRaw = strtol(*args, &tail, 10);
-
-    if (tail != *args && isWhiteSpace(*tail))
-    {
-        *(tail++) = '\0';
-    }
-    else if (tail && *tail)                                 // some not whitespace symbol
-    {
-        return false;                                        // args not modified and can be re-parsed
-    }
-
-    if (valRaw < std::numeric_limits<int32>::min() || valRaw > std::numeric_limits<int32>::max())
-    {
-        return false;
-    }
-
-    // value successfully extracted
-    val = int32(valRaw);
-    *args = tail;
-    return true;
-}
-
-/**
- * Function extract to val arg optional signed integer value or use default value. Fail if extracted not signed integer.
- *
- * @param args    variable pointer to non parsed args string, updated at function call to new position (with skipped white spaces)
- * @param val     return extracted value if function success, in fail case original value unmodified
- * @param defVal  default value used if no data for extraction in args
- * @return        true if value extraction successful
- */
-bool  ChatHandler::ExtractOptInt32(char** args, int32& val, int32 defVal)
-{
-    if (!*args || !** args)
-    {
-        val = defVal;
-        return true;
-    }
-
-    return ExtractInt32(args, val);
-}
-
-/**
- * Function extract to val arg unsigned integer value or fail
- *
- * @param args variable pointer to non parsed args string, updated at function call to new position (with skipped white spaces)
- * @param val  return extracted value if function success, in fail case original value unmodified
- * @param base set used base for extracted value format (10 for decimal, 16 for hex, etc), 0 let auto select by system internal function
- * @return     true if value extraction successful
- */
-bool  ChatHandler::ExtractUInt32Base(char** args, uint32& val, uint32 base)
-{
-    if (!*args || !** args)
-    {
-        return false;
-    }
-
-    char* tail = *args;
-
-    unsigned long valRaw = strtoul(*args, &tail, base);
-
-    if (tail != *args && isWhiteSpace(*tail))
-    {
-        *(tail++) = '\0';
-    }
-    else if (tail && *tail)                                 // some not whitespace symbol
-    {
-        return false;                                        // args not modified and can be re-parsed
-    }
-
-    if (valRaw > std::numeric_limits<uint32>::max())
-    {
-        return false;
-    }
-
-    // value successfully extracted
-    val = uint32(valRaw);
-    *args = tail;
-
-    SkipWhiteSpaces(args);
-    return true;
-}
-
-/**
- * Function extract to val arg optional unsigned integer value or use default value. Fail if extracted not unsigned integer.
- *
- * @param args    variable pointer to non parsed args string, updated at function call to new position (with skipped white spaces)
- * @param val     return extracted value if function success, in fail case original value unmodified
- * @param defVal  default value used if no data for extraction in args
- * @return        true if value extraction successful
- */
-bool  ChatHandler::ExtractOptUInt32(char** args, uint32& val, uint32 defVal)
-{
-    if (!*args || !** args)
-    {
-        val = defVal;
-        return true;
-    }
-
-    return ExtractUInt32(args, val);
-}
-
-/**
- * Function extract to val arg float value or fail
- *
- * @param args variable pointer to non parsed args string, updated at function call to new position (with skipped white spaces)
- * @param val  return extracted value if function success, in fail case original value unmodified
- * @return     true if value extraction successful
- */
-bool  ChatHandler::ExtractFloat(char** args, float& val)
-{
-    if (!*args || !** args)
-    {
-        return false;
-    }
-
-    char* tail = *args;
-
-    double valRaw = strtod(*args, &tail);
-
-    if (tail != *args && isWhiteSpace(*tail))
-    {
-        *(tail++) = '\0';
-    }
-    else if (tail && *tail)                                 // some not whitespace symbol
-    {
-        return false;                                        // args not modified and can be re-parsed
-    }
-
-    // value successfully extracted
-    val = float(valRaw);
-    *args = tail;
-
-    SkipWhiteSpaces(args);
-    return true;
-}
-
-/**
- * Function extract to val arg optional float value or use default value. Fail if extracted not float.
- *
- * @param args    variable pointer to non parsed args string, updated at function call to new position (with skipped white spaces)
- * @param val     return extracted value if function success, in fail case original value unmodified
- * @param defVal  default value used if no data for extraction in args
- * @return        true if value extraction successful
- */
-bool  ChatHandler::ExtractOptFloat(char** args, float& val, float defVal)
-{
-    if (!*args || !** args)
-    {
-        val = defVal;
-        return true;
-    }
-
-    return ExtractFloat(args, val);
-}
-
-/**
- * Function extract name-like string (from non-numeric or special symbol until whitespace)
- *
- * @param args variable pointer to non parsed args string, updated at function call to new position (with skipped white spaces)
- * @param lit  optional explicit literal requirement. function fail if literal is not starting substring of lit.
- *             Note: function in same way fail if no any literal or literal not fit in this case. Need additional check for select specific fail case
- * @return     name/number-like string without whitespaces, or NULL if args empty or not appropriate content.
- */
-char* ChatHandler::ExtractLiteralArg(char** args, char const* lit /*= NULL*/)
-{
-    if (!*args || !** args)
-    {
-        return NULL;
-    }
-
-    char* head = *args;
-
-    // reject quoted string or link (|-started text)
-    switch (head[0])
-    {
-            // reject quoted string
-        case '[': case '\'': case '"':
-            return NULL;
-            // reject link (|-started text)
-        case '|':
-            // client replace all | by || in raw text
-            if (head[1] != '|')
-            {
-                return NULL;
-            }
-            ++head;                                         // skip one |
-            break;
-        default: break;
-    }
-
-    if (lit)
-    {
-        int l = strlen(lit);
-
-        int largs = 0;
-        while (head[largs] && !isWhiteSpace(head[largs]))
-        {
-            ++largs;
-        }
-
-        if (largs < l)
-        {
-            l = largs;
-        }
-
-        int diff = strncmp(head, lit, l);
-
-        if (diff != 0)
-        {
-            return NULL;
-        }
-
-        if (head[l] && !isWhiteSpace(head[l]))
-        {
-            return NULL;
-        }
-
-        char* arg = head;
-
-        if (head[l])
-        {
-            head[l] = '\0';
-
-            head += l + 1;
-
-            *args = head;
-        }
-        else
-        {
-            *args = head + l;
-        }
-
-        SkipWhiteSpaces(args);
-        return arg;
-    }
-
-    char* name = strtok(head, " ");
-
-    char* tail = strtok(NULL, "");
-
-    *args = tail ? tail : (char*)"";                        // *args don't must be NULL
-
-    SkipWhiteSpaces(args);
-
-    return name;
-}
-
-/**
- * Function extract quote-like string (any characters guarded by some special character, in our cases ['")
- *
- * @param args variable pointer to non parsed args string, updated at function call to new position (with skipped white spaces)
- * @param asis control save quote string wrappers
- * @return     quote-like string, or NULL if args empty or not appropriate content.
- */
-char* ChatHandler::ExtractQuotedArg(char** args, bool asis /*= false*/)
-{
-    if (!*args || !** args)
-    {
-        return NULL;
-    }
-
-    if (**args != '\'' &&**  args != '"' &&**  args != '[')
-    {
-        return NULL;
-    }
-
-    char guard = (*args)[0];
-
-    if (guard == '[')
-    {
-        guard = ']';
-    }
-
-    char* tail = (*args) + 1;                               // start scan after first quote symbol
-    char* head = asis ? *args : tail;                       // start arg
-
-    while (*tail && *tail != guard)
-    {
-        ++tail;
-    }
-
-    if (!*tail || (tail[1] && !isWhiteSpace(tail[1])))      // fail
-    {
-        return NULL;
-    }
-
-    if (!tail[1])                                           // quote is last char in string
-    {
-        if (!asis)
-        {
-            *tail = '\0';
-        }
-    }
-    else                                                    // quote isn't last char
-    {
-        if (asis)
-        {
-            ++tail;
-        }
-
-        *tail = '\0';
-    }
-
-    *args = tail + 1;
-
-    SkipWhiteSpaces(args);
-
-    return head;
-}
-
-/**
- * Function extract quote-like string or literal if quote not detected
- *
- * @param args variable pointer to non parsed args string, updated at function call to new position (with skipped white spaces)
- * @param asis control save quote string wrappers
- * @return     quote/literal string, or NULL if args empty or not appropriate content.
- */
-char* ChatHandler::ExtractQuotedOrLiteralArg(char** args, bool asis /*= false*/)
-{
-    char* arg = ExtractQuotedArg(args, asis);
-    if (!arg)
-    {
-        arg = ExtractLiteralArg(args);
-    }
-    return arg;
-}
-
-/**
- * Function extract on/off literals as boolean values
- *
- * @param args variable pointer to non parsed args string, updated at function call to new position (with skipped white spaces)
- * @param val  return extracted value if function success, in fail case original value unmodified
- * @return     true at success
- */
-bool  ChatHandler::ExtractOnOff(char** args, bool& value)
-{
-    char* arg = ExtractLiteralArg(args);
-    if (!arg)
-    {
-        return false;
-    }
-
-    if (strncmp(arg, "on", 3) == 0)
-    {
-        value = true;
-    }
-    else if (strncmp(arg, "off", 4) == 0)
-    {
-        value = false;
-    }
-    else
-    {
-        return false;
-    }
-
-    return true;
-}
-
-/**
- * Function extract shift-link-like string (any characters guarded by | and |h|r with some additional internal structure check)
- *
- * @param args variable pointer to non parsed args string, updated at function call to new position (with skipped white spaces)
- *
- * @param linkTypes  optional NULL-terminated array of link types, shift-link must fit one from link type from array if provided or extraction fail
- *
- * @param found_idx  if not NULL then at return index in linkTypes that fit shift-link type, if extraction fail then non modified
- *
- * @param keyPair    if not NULL then pointer to 2-elements array for return start and end pointer for found key
- *                   if extraction fail then non modified
- *
- * @param somethingPair then pointer to 2-elements array for return start and end pointer if found.
- *                   if not NULL then shift-link must have data field, if extraction fail then non modified
- *
- * @return     shift-link-like string, or NULL if args empty or not appropriate content.
- */
-char* ChatHandler::ExtractLinkArg(char** args, char const* const* linkTypes /*= NULL*/, int* foundIdx /*= NULL*/, char** keyPair /*= NULL*/, char** somethingPair /*= NULL*/)
-{
-    if (!*args || !** args)
-    {
-        return NULL;
-    }
-
-    // skip if not linked started or encoded single | (doubled by client)
-    if ((*args)[0] != '|' || (*args)[1] == '|')
-    {
-        return NULL;
-    }
-
-    // |color|Hlinktype:key:data...|h[name]|h|r
-
-    char* head = *args;
-
-    // [name] Shift-click form |color|linkType:key|h[name]|h|r
-    // or
-    // [name] Shift-click form |color|linkType:key:something1:...:somethingN|h[name]|h|r
-    // or
-    // [name] Shift-click form |linkType:key|h[name]|h|r
-
-    // |color|Hlinktype:key:data...|h[name]|h|r
-
-    char* tail = (*args) + 1;                               // skip |
-
-    if (*tail != 'H')                                       // skip color part, some links can not have color part
-    {
-        while (*tail && *tail != '|')
-        {
-            ++tail;
-        }
-
-        if (!*tail)
-        {
-            return NULL;
-        }
-
-        // |Hlinktype:key:data...|h[name]|h|r
-
-        ++tail;                                             // skip |
-    }
-
-    // Hlinktype:key:data...|h[name]|h|r
-
-    if (*tail != 'H')
-    {
-        return NULL;
-    }
-
-    int linktype_idx = 0;
-
-    if (linkTypes)                                          // check link type if provided
-    {
-        // check linktypes (its include H in name)
-        for (; linkTypes[linktype_idx]; ++linktype_idx)
-        {
-            // exactly string with follow : or |
-            int l = strlen(linkTypes[linktype_idx]);
-            if (strncmp(tail, linkTypes[linktype_idx], l) == 0 &&
-                (tail[l] == ':' || tail[l] == '|'))
-            {
-                break;
-            }
-        }
-
-        // is search fail?
-        if (!linkTypes[linktype_idx])                       // NULL terminator in last element
-        {
-            return NULL;
-        }
-
-        tail += strlen(linkTypes[linktype_idx]);            // skip linktype string
-
-        // :key:data...|h[name]|h|r
-
-        if (*tail != ':')
-        {
-            return NULL;
-        }
-    }
-    else
-    {
-        while (*tail && *tail != ':')                       // skip linktype string
-        {
-            ++tail;
-        }
-
-        if (!*tail)
-        {
-            return NULL;
-        }
-    }
-
-    ++tail;
-
-    // key:data...|h[name]|h|r
-    char* keyStart = tail;                                  // remember key start for return
-
-    while (*tail && *tail != '|' && *tail != ':')
-    {
-        ++tail;
-    }
-
-    if (!*tail)
-    {
-        return NULL;
-    }
-
-    char* keyEnd = tail;                                    // remember key end for truncate
-
-    // |h[name]|h|r or :something...|h[name]|h|r
-
-    char* somethingStart = tail + 1;
-    char* somethingEnd   = tail + 1;                        // will updated later if need
-
-    if (*tail == ':' && somethingPair)                      // optional data extraction
-    {
-        // :something...|h[name]|h|r
-        ++tail;
-
-        // something|h[name]|h|r or something:something2...|h[name]|h|r
-
-        while (*tail && *tail != '|' && *tail != ':')
-        {
-            ++tail;
-        }
-
-        if (!*tail)
-        {
-            return NULL;
-        }
-
-        somethingEnd = tail;                                // remember data end for truncate
-    }
-
-    // |h[name]|h|r or :something2...|h[name]|h|r
-
-    while (*tail && (*tail != '|' || *(tail + 1) != 'h'))   // skip ... part if exist
-    {
-        ++tail;
-    }
-
-    if (!*tail)
-    {
-        return NULL;
-    }
-
-    // |h[name]|h|r
-
-    tail += 2;                                              // skip |h
-
-    // [name]|h|r
-    if (!*tail || *tail != '[')
-    {
-        return NULL;
-    }
-
-    while (*tail && (*tail != ']' || *(tail + 1) != '|'))   // skip name part
-    {
-        ++tail;
-    }
-
-    tail += 2;                                              // skip ]|
-
-    // h|r
-    if (!*tail || *tail != 'h'  || *(tail + 1) != '|')
-    {
-        return NULL;
-    }
-
-    tail += 2;                                              // skip h|
-
-    // r
-    if (!*tail || *tail != 'r' || (*(tail + 1) && !isWhiteSpace(*(tail + 1))))
-    {
-        return NULL;
-    }
-
-    ++tail;                                                 // skip r
-
-    // success
-
-    if (*tail)                                              // truncate all link string
-    {
-        *(tail++) = '\0';
-    }
-
-    if (foundIdx)
-    {
-        *foundIdx = linktype_idx;
-    }
-
-    if (keyPair)
-    {
-        keyPair[0] = keyStart;
-        keyPair[1] = keyEnd;
-    }
-
-    if (somethingPair)
-    {
-        somethingPair[0] = somethingStart;
-        somethingPair[1] = somethingEnd;
-    }
-
-    *args = tail;
-
-    SkipWhiteSpaces(args);
-
-    return head;
-}
-
-/**
- * Function extract name/number/quote/shift-link-like string
- *
- * @param args variable pointer to non parsed args string, updated at function call to new position (with skipped white spaces)
- * @param asis control save quote string wrappers
- * @return     extracted arg string, or NULL if args empty or not appropriate content.
- */
-char* ChatHandler::ExtractArg(char** args, bool asis /*= false*/)
-{
-    if (!*args || !** args)
-    {
-        return NULL;
-    }
-
-    char* arg = ExtractQuotedOrLiteralArg(args, asis);
-    if (!arg)
-    {
-        arg = ExtractLinkArg(args);
-    }
-
-    return arg;
-}
-
-/**
- * Function extract name/quote/number/shift-link-like string, and return it if args have more non-whitespace data
- *
- * @param args variable pointer to non parsed args string, updated at function call to new position (with skipped white spaces)
- *             if args have only single arg then args still pointing to this arg (unmodified pointer)
- * @return     extracted string, or NULL if args empty or not appropriate content or have single arg totally.
- */
-char* ChatHandler::ExtractOptNotLastArg(char** args)
-{
-    char* arg = ExtractArg(args, true);
-
-    // have more data
-    if (*args &&**  args)
-    {
-        return arg;
-    }
-
-    // optional name not found
-    *args = arg ? arg : (char*)"";                          // *args don't must be NULL
-
-    return NULL;
-}
-
-/**
- * Function extract data from shift-link "|color|LINKTYPE:RETURN:SOMETHING1|h[name]|h|r if linkType == LINKTYPE
- * It also extract literal/quote if not shift-link in args
- *
- * @param args       variable pointer to non parsed args string, updated at function call to new position (with skipped white spaces)
- *                   if args have sift link with linkType != LINKTYPE then args still pointing to this arg (unmodified pointer)
- *
- * @param linkType   shift-link must fit by link type to this arg value or extraction fail
- *
- * @param something1 if not NULL then shift-link must have data field and it returned into this arg
- *                   if extraction fail then non modified
- *
- * @return           extracted key, or NULL if args empty or not appropriate content or not fit to linkType.
- */
-char* ChatHandler::ExtractKeyFromLink(char** text, char const* linkType, char** something1 /*= NULL*/)
-{
-    char const* linkTypes[2];
-    linkTypes[0] = linkType;
-    linkTypes[1] = NULL;
-
-    int foundIdx;
-
-    return ExtractKeyFromLink(text, linkTypes, &foundIdx, something1);
-}
-
-/**
- * Function extract data from shift-link "|color|LINKTYPE:RETURN:SOMETHING1|h[name]|h|r if LINKTYPE in linkTypes array
- * It also extract literal/quote if not shift-link in args
- *
- * @param args       variable pointer to non parsed args string, updated at function call to new position (with skipped white spaces)
- *                   if args have sift link with linkType != LINKTYPE then args still pointing to this arg (unmodified pointer)
- *
- * @param linkTypes  NULL-terminated array of link types, shift-link must fit one from link type from array or extraction fail
- *
- * @param found_idx  if not NULL then at return index in linkTypes that fit shift-link type, for non-link case return -1
- *                   if extraction fail then non modified
- *
- * @param something1 if not NULL then shift-link must have data field and it returned into this arg
- *                   if extraction fail then non modified
- *
- * @return           extracted key, or NULL if args empty or not appropriate content or not fit to linkType.
- */
-char* ChatHandler::ExtractKeyFromLink(char** text, char const* const* linkTypes, int* found_idx, char** something1 /*= NULL*/)
-{
-    // skip empty
-    if (!*text || !** text)
-    {
-        return NULL;
-    }
-
-    // return non link case
-    char* arg = ExtractQuotedOrLiteralArg(text);
-    if (arg)
-    {
-        if (found_idx)
-        {
-            *found_idx = -1;                                 // special index case
-        }
-
-        return arg;
-    }
-
-    char* keyPair[2];
-    char* somethingPair[2];
-
-    arg = ExtractLinkArg(text, linkTypes, found_idx, keyPair, something1 ? somethingPair : NULL);
-    if (!arg)
-    {
-        return NULL;
-    }
-
-    *keyPair[1] = '\0';                                     // truncate key string
-
-    if (something1)
-    {
-        *somethingPair[1] = '\0';                           // truncate data string
-        *something1 = somethingPair[0];
-    }
-
-    return keyPair[0];
-}
-
-/**
- * Function extract uint32 key from shift-link "|color|LINKTYPE:RETURN|h[name]|h|r if linkType == LINKTYPE
- * It also extract direct number if not shift-link in text
- *
- * @param text       variable pointer to non parsed argument string, updated at function call to new position (with skipped white spaces)
- *                   if text has shift link with linkType != LINKTYPE then text still points to this argument (unmodified pointer)
- *
- * @param linkType   shift-link must fit by link type to this arg value or extraction fail
- *
- * @param value      store result value at success return, not modified at fail
- *
- * @return           true if extraction succesful
- */
-bool ChatHandler::ExtractUint32KeyFromLink(char** text, char const* linkType, uint32& value)
-{
-    char* arg = ExtractKeyFromLink(text, linkType);
-    if (!arg)
-    {
-        return false;
-    }
-
-    return ExtractUInt32(&arg, value);
-}
-
-/**
- * @brief Retrieves a game object on the current map by low GUID and entry.
- *
- * @param lowguid The game object low GUID.
- * @param entry The game object entry identifier.
- * @return GameObject* The matching game object, or NULL if not found.
- */
-GameObject* ChatHandler::GetGameObjectWithGuid(uint32 lowguid, uint32 entry)
-{
-    if (!m_session)
-    {
-        return NULL;
-    }
-
-    Player* pl = m_session->GetPlayer();
-
-    return pl->GetMap()->GetGameObject(ObjectGuid(HIGHGUID_GAMEOBJECT, entry, lowguid));
-}
-
-enum SpellLinkType
-{
-    SPELL_LINK_RAW     = -1,                                // non-link case
-    SPELL_LINK_SPELL   = 0,
-    SPELL_LINK_TALENT  = 1,
-    SPELL_LINK_ENCHANT = 2,
-};
-
-static char const* const spellKeys[] =
-{
-    "Hspell",                                               // normal spell
-    "Htalent",                                              // talent spell
-    "Henchant",                                             // enchanting recipe spell
-    NULL
-};
-
-/**
- * @brief Extracts a spell id from a raw argument or supported spell-related link.
- *
- * @param text The argument text pointer to parse.
- * @return uint32 The extracted spell id, or 0 on failure.
- */
-uint32 ChatHandler::ExtractSpellIdFromLink(char** text)
-{
-    // number or [name] Shift-click form |color|Henchant:recipe_spell_id|h[prof_name: recipe_name]|h|r
-    // number or [name] Shift-click form |color|Hspell:spell_id|h[name]|h|r
-    // number or [name] Shift-click form |color|Htalent:talent_id,rank|h[name]|h|r
-    int type;
-    char* param1_str = NULL;
-    char* idS = ExtractKeyFromLink(text, spellKeys, &type, &param1_str);
-    if (idS)
-    {
-        uint32 id;
-        if (ExtractUInt32(&idS, id))
-        {
-            switch (type)
-            {
-                case SPELL_LINK_RAW:
-                case SPELL_LINK_SPELL:
-                case SPELL_LINK_ENCHANT:
-                    return id;
-                case SPELL_LINK_TALENT:
-                {
-                    // talent
-                    TalentEntry const* talentEntry = sTalentStore.LookupEntry(id);
-                    int32 rank;
-                    if (talentEntry && ExtractInt32(&param1_str, rank))
-                    {
-                        if (rank < 0) // unlearned talent have in shift-link field -1 as rank
-                        {
-                            rank = 0;
-                        }
-                        if (rank < MAX_TALENT_RANK)
-                        {
-                            return talentEntry->RankID[rank];
-                        }
-                    }
-                    break;
-                }
-            }
-        }
-    }
-
-    // Name-based fallback
-    char const* lookupName = idS ? idS : *text;
-    if (!lookupName || !*lookupName)
-    {
-        return 0;
-    }
-
-    LocaleConstant locale = GetSessionDbcLocale();
-
-    std::wstring wname;
-    if (!Utf8toWStr(lookupName, wname))
-    {
-        return 0;
-    }
-    wstrToLower(wname);
-
-    uint32 exactId = 0;
-    std::vector<std::pair<uint32, std::string>> candidates;
-    std::vector<std::pair<uint32, std::string>> fuzzyCandidates;
-
-    for (uint32 id = 0; id < sSpellStore.GetNumRows(); ++id)
-    {
-        SpellEntry const* spell = sSpellStore.LookupEntry(id);
-        if (!spell)
-        {
-            continue;
-        }
-        std::string sName = spell->SpellName[locale];
-
-        if (spell->Effect[EFFECT_INDEX_0] == SPELL_EFFECT_LEARN_SPELL ||
-            spell->Effect[EFFECT_INDEX_1] == SPELL_EFFECT_LEARN_SPELL ||
-            spell->Effect[EFFECT_INDEX_2] == SPELL_EFFECT_LEARN_SPELL ||
-            sName.empty() || sName.find("Test ") == 0 ||
-            sName.find("ZZ") == 0 || sName.find("zz") == 0)
-        {
-            continue;
-        }
-
-        std::wstring wSpellName;
-        if (!Utf8toWStr(sName, wSpellName))
-        {
-            continue;
-        }
-        wstrToLower(wSpellName);
-
-        if (wSpellName == wname)
-        {
-            if (exactId == 0)
-            {
-                exactId = id;
-            }
-            candidates.push_back({id, spell->SpellName[locale]});
-        }
-        else if (exactId == 0)
-        {
-            int loc = locale;
-            if (!Utf8FitTo(sName, wname))
-            {
-                loc = 0;
-                for (; loc < MAX_LOCALE; ++loc)
-                {
-                    if (loc == locale)
-                    {
-                        continue;
-                    }
-                    sName = spell->SpellName[loc];
-                    if (sName.empty())
-                    {
-                        continue;
-                    }
-                    if (Utf8FitTo(sName, wname))
-                    {
-                        break;
-                    }
-                }
-            }
-
-            if (loc < MAX_LOCALE)
-            {
-                fuzzyCandidates.push_back({id, spell->SpellName[locale]});
-            }
-        }
-    }
-
-    if (!candidates.empty())
-    {
-        if (candidates.size() == 1)
-        {
-            return exactId;
-        }
-
-        SendSysMessage("Multiple spells found with that name:");
-        for (auto const& c : candidates)
-        {
-            PSendSysMessage("  %u - %s", c.first, c.second.c_str());
-        }
-        SetSentErrorMessage(true);
-        return 0;
-    }
-
-    if (fuzzyCandidates.empty())
-    {
-        SendSysMessage(LANG_COMMAND_NOSPELLFOUND);
-        SetSentErrorMessage(true);
-        return 0;
-    }
-
-    SendSysMessage("No exact match. Close matches:");
-    for (auto const& c : fuzzyCandidates)
-    {
-        PSendSysMessage("  %u - %s", c.first, c.second.c_str());
-    }
-    SetSentErrorMessage(true);
-    return 0;
-}
-
-/**
- * @brief Extracts a game teleport definition from a raw argument or teleport link.
- *
- * @param text The argument text pointer to parse.
- * @return GameTele const* The matching teleport definition, or NULL on failure.
- */
-GameTele const* ChatHandler::ExtractGameTeleFromLink(char** text)
-{
-    // id, or string, or [name] Shift-click form |color|Htele:id|h[name]|h|r
-    char* cId = ExtractKeyFromLink(text, "Htele");
-    if (!cId)
-    {
-        return NULL;
-    }
-
-    // id case (explicit or from shift link)
-    uint32 id;
-    if (ExtractUInt32(&cId, id))
-    {
-        return sObjectMgr.GetGameTele(id);
-    }
-    else
-    {
-        return sObjectMgr.GetGameTele(cId);
-    }
-}
-
-enum GuidLinkType
-{
-    GUID_LINK_RAW        = -1,                              // non-link case
-    GUID_LINK_PLAYER     = 0,
-    GUID_LINK_CREATURE   = 1,
-    GUID_LINK_GAMEOBJECT = 2,
-};
-
-static char const* const guidKeys[] =
-{
-    "Hplayer",
-    "Hcreature",
-    "Hgameobject",
-    NULL
-};
-
-/**
- * @brief Extracts a player, creature, or game object GUID from a raw argument or link.
- *
- * @param text The argument text pointer to parse.
- * @return ObjectGuid The extracted GUID, or an empty GUID on failure.
- */
-ObjectGuid ChatHandler::ExtractGuidFromLink(char** text)
-{
-    int type = 0;
-
-    // |color|Hcreature:creature_guid|h[name]|h|r
-    // |color|Hgameobject:go_guid|h[name]|h|r
-    // |color|Hplayer:name|h[name]|h|r
-    char* idS = ExtractKeyFromLink(text, guidKeys, &type);
-    if (!idS)
-    {
-        return ObjectGuid();
-    }
-
-    switch (type)
-    {
-        case GUID_LINK_RAW:
-        case GUID_LINK_PLAYER:
-        {
-            std::string name = idS;
-            if (!normalizePlayerName(name))
-            {
-                return ObjectGuid();
-            }
-
-            if (Player* player = sObjectMgr.GetPlayer(name.c_str()))
-            {
-                return player->GetObjectGuid();
-            }
-
-            return sObjectMgr.GetPlayerGuidByName(name);
-        }
-        case GUID_LINK_CREATURE:
-        {
-            uint32 lowguid;
-            if (!ExtractUInt32(&idS, lowguid))
-            {
-                return ObjectGuid();
-            }
-
-            if (CreatureData const* data = sObjectMgr.GetCreatureData(lowguid))
-            {
-                return data->GetObjectGuid(lowguid);
-            }
-            else
-            {
-                return ObjectGuid();
-            }
-        }
-        case GUID_LINK_GAMEOBJECT:
-        {
-            uint32 lowguid;
-            if (!ExtractUInt32(&idS, lowguid))
-            {
-                return ObjectGuid();
-            }
-
-            if (GameObjectData const* data = sObjectMgr.GetGOData(lowguid))
-            {
-                return ObjectGuid(HIGHGUID_GAMEOBJECT, data->id, lowguid);
-            }
-            else
-            {
-                return ObjectGuid();
-            }
-        }
-    }
-
-    // unknown type?
-    return ObjectGuid();
-}
-
-enum LocationLinkType
-{
-    LOCATION_LINK_RAW               = -1,                   // non-link case
-    LOCATION_LINK_PLAYER            = 0,
-    LOCATION_LINK_TELE              = 1,
-    LOCATION_LINK_TAXINODE          = 2,
-    LOCATION_LINK_CREATURE          = 3,
-    LOCATION_LINK_GAMEOBJECT        = 4,
-    LOCATION_LINK_CREATURE_ENTRY    = 5,
-    LOCATION_LINK_GAMEOBJECT_ENTRY  = 6,
-    LOCATION_LINK_AREATRIGGER       = 7,
-    LOCATION_LINK_AREATRIGGER_TARGET = 8,
-};
-
-static char const* const locationKeys[] =
-{
-    "Htele",
-    "Htaxinode",
-    "Hplayer",
-    "Hcreature",
-    "Hgameobject",
-    "Hcreature_entry",
-    "Hgameobject_entry",
-    "Hareatrigger",
-    "Hareatrigger_target",
-    NULL
-};
-
-/**
- * @brief Extracts map coordinates from a supported location-like link or player reference.
- *
- * @param text The argument text pointer to parse.
- * @param mapid Receives the destination map id.
- * @param x Receives the destination X coordinate.
- * @param y Receives the destination Y coordinate.
- * @param z Receives the destination Z coordinate.
- * @return true if a location was extracted; otherwise false.
- */
-bool ChatHandler::ExtractLocationFromLink(char** text, uint32& mapid, float& x, float& y, float& z)
-{
-    int type = 0;
-
-    // |color|Hplayer:name|h[name]|h|r
-    // |color|Htele:id|h[name]|h|r
-    // |color|Htaxinode:id|h[name]|h|r
-    // |color|Hcreature:creature_guid|h[name]|h|r
-    // |color|Hgameobject:go_guid|h[name]|h|r
-    // |color|Hcreature_entry:creature_id|h[name]|h|r
-    // |color|Hgameobject_entry:go_id|h[name]|h|r
-    // |color|Hareatrigger:id|h[name]|h|r
-    // |color|Hareatrigger_target:id|h[name]|h|r
-    char* idS = ExtractKeyFromLink(text, locationKeys, &type);
-    if (!idS)
-    {
-        return false;
-    }
-
-    switch (type)
-    {
-        case LOCATION_LINK_RAW:
-        case LOCATION_LINK_PLAYER:
-        {
-            std::string name = idS;
-            if (!normalizePlayerName(name))
-            {
-                return false;
-            }
-
-            if (Player* player = sObjectMgr.GetPlayer(name.c_str()))
-            {
-                mapid = player->GetMapId();
-                x = player->GetPositionX();
-                y = player->GetPositionY();
-                z = player->GetPositionZ();
-                return true;
-            }
-
-            if (ObjectGuid guid = sObjectMgr.GetPlayerGuidByName(name))
-            {
-                // to point where player stay (if loaded)
-                float o;
-                bool in_flight;
-                return Player::LoadPositionFromDB(guid, mapid, x, y, z, o, in_flight);
-            }
-
-            return false;
-        }
-        case LOCATION_LINK_TELE:
-        {
-            uint32 id;
-            if (!ExtractUInt32(&idS, id))
-            {
-                return false;
-            }
-
-            GameTele const* tele = sObjectMgr.GetGameTele(id);
-            if (!tele)
-            {
-                return false;
-            }
-            mapid = tele->mapId;
-            x = tele->position_x;
-            y = tele->position_y;
-            z = tele->position_z;
-            return true;
-        }
-        case LOCATION_LINK_TAXINODE:
-        {
-            uint32 id;
-            if (!ExtractUInt32(&idS, id))
-            {
-                return false;
-            }
-
-            TaxiNodesEntry const* node = sTaxiNodesStore.LookupEntry(id);
-            if (!node)
-            {
-                return false;
-            }
-            mapid = node->map_id;
-            x = node->x;
-            y = node->y;
-            z = node->z;
-            return true;
-        }
-        case LOCATION_LINK_CREATURE:
-        {
-            uint32 lowguid;
-            if (!ExtractUInt32(&idS, lowguid))
-            {
-                return false;
-            }
-
-            if (CreatureData const* data = sObjectMgr.GetCreatureData(lowguid))
-            {
-                mapid = data->mapid;
-                x = data->posX;
-                y = data->posY;
-                z = data->posZ;
-                return true;
-            }
-            else
-            {
-                return false;
-            }
-        }
-        case LOCATION_LINK_GAMEOBJECT:
-        {
-            uint32 lowguid;
-            if (!ExtractUInt32(&idS, lowguid))
-            {
-                return false;
-            }
-
-            if (GameObjectData const* data = sObjectMgr.GetGOData(lowguid))
-            {
-                mapid = data->mapid;
-                x = data->posX;
-                y = data->posY;
-                z = data->posZ;
-                return true;
-            }
-            else
-            {
-                return false;
-            }
-        }
-        case LOCATION_LINK_CREATURE_ENTRY:
-        {
-            uint32 id;
-            if (!ExtractUInt32(&idS, id))
-            {
-                return false;
-            }
-
-            if (ObjectMgr::GetCreatureTemplate(id))
-            {
-                FindCreatureData worker(id, m_session ? m_session->GetPlayer() : NULL);
-
-                sObjectMgr.DoCreatureData(worker);
-
-                if (CreatureDataPair const* dataPair = worker.GetResult())
-                {
-                    mapid = dataPair->second.mapid;
-                    x = dataPair->second.posX;
-                    y = dataPair->second.posY;
-                    z = dataPair->second.posZ;
-                    return true;
-                }
-                else
-                {
-                    return false;
-                }
-            }
-            else
-            {
-                return false;
-            }
-        }
-        case LOCATION_LINK_GAMEOBJECT_ENTRY:
-        {
-            uint32 id;
-            if (!ExtractUInt32(&idS, id))
-            {
-                return false;
-            }
-
-            if (ObjectMgr::GetGameObjectInfo(id))
-            {
-                FindGOData worker(id, m_session ? m_session->GetPlayer() : NULL);
-
-                sObjectMgr.DoGOData(worker);
-
-                if (GameObjectDataPair const* dataPair = worker.GetResult())
-                {
-                    mapid = dataPair->second.mapid;
-                    x = dataPair->second.posX;
-                    y = dataPair->second.posY;
-                    z = dataPair->second.posZ;
-                    return true;
-                }
-                else
-                {
-                    return false;
-                }
-            }
-            else
-            {
-                return false;
-            }
-        }
-        case LOCATION_LINK_AREATRIGGER:
-        {
-            uint32 id;
-            if (!ExtractUInt32(&idS, id))
-            {
-                return false;
-            }
-
-            AreaTriggerEntry const* atEntry = sAreaTriggerStore.LookupEntry(id);
-            if (!atEntry)
-            {
-                PSendSysMessage(LANG_COMMAND_GOAREATRNOTFOUND, id);
-                SetSentErrorMessage(true);
-                return false;
-            }
-
-            mapid = atEntry->mapid;
-            x = atEntry->x;
-            y = atEntry->y;
-            z = atEntry->z;
-            return true;
-        }
-        case LOCATION_LINK_AREATRIGGER_TARGET:
-        {
-            uint32 id;
-            if (!ExtractUInt32(&idS, id))
-            {
-                return false;
-            }
-
-            if (!sAreaTriggerStore.LookupEntry(id))
-            {
-                PSendSysMessage(LANG_COMMAND_GOAREATRNOTFOUND, id);
-                SetSentErrorMessage(true);
-                return false;
-            }
-
-            AreaTrigger const* at = sObjectMgr.GetAreaTrigger(id);
-            if (!at)
-            {
-                PSendSysMessage(LANG_AREATRIGER_NOT_HAS_TARGET, id);
-                SetSentErrorMessage(true);
-                return false;
-            }
-
-            mapid = at->target_mapId;
-            x = at->target_X;
-            y = at->target_Y;
-            z = at->target_Z;
-            return true;
-        }
-    }
-
-    // unknown type?
-    return false;
-}
-
-/**
- * @brief Extracts and normalizes a player name from a player link.
- *
- * @param text The argument text pointer to parse.
- * @return std::string The normalized player name, or an empty string on failure.
- */
-std::string ChatHandler::ExtractPlayerNameFromLink(char** text)
-{
-    // |color|Hplayer:name|h[name]|h|r
-    char* name_str = ExtractKeyFromLink(text, "Hplayer");
-    if (!name_str)
-    {
-        return "";
-    }
-
-    std::string name = name_str;
-    if (!normalizePlayerName(name))
-    {
-        return "";
-    }
-
-    return name;
-}
-
-/**
- * Function extract at least one from request player data (pointer/guid/name) from args name/shift-link or selected player if no args
- *
- * @param args        variable pointer to non parsed args string, updated at function call to new position (with skipped white spaces)
- *
- * @param player      optional arg   One from 3 optional args must be provided at least (or more).
- * @param player_guid optional arg   For function success only one from provided args need get result
- * @param player_name optional arg   But if early arg get value then all later args will have its (if requested)
- *                                   if player_guid requested and not found then name also will not found
- *                                   So at success can be returned 2 cases: (player/guid/name) or (guid/name)
- *
- * @return           true if extraction successful
- */
-bool ChatHandler::ExtractPlayerTarget(char** args, Player** player /*= NULL*/, ObjectGuid* player_guid /*= NULL*/, std::string* player_name /*= NULL*/)
-{
-    if (*args &&**  args)
-    {
-        std::string name = ExtractPlayerNameFromLink(args);
-        if (name.empty())
-        {
-            SendSysMessage(LANG_PLAYER_NOT_FOUND);
-            SetSentErrorMessage(true);
-            return false;
-        }
-
-        Player* pl = sObjectMgr.GetPlayer(name.c_str());
-
-        // if allowed player pointer
-        if (player)
-        {
-            *player = pl;
-        }
-
-        // if need guid value from DB (in name case for check player existence)
-        ObjectGuid guid = !pl && (player_guid || player_name) ? sObjectMgr.GetPlayerGuidByName(name) : ObjectGuid();
-
-        // if allowed player guid (if no then only online players allowed)
-        if (player_guid)
-        {
-            *player_guid = pl ? pl->GetObjectGuid() : guid;
-        }
-
-        if (player_name)
-        {
-            *player_name = pl || guid ? name : "";
-        }
-    }
-    else
-    {
-        Player* pl = getSelectedPlayer();
-        // if allowed player pointer
-        if (player)
-        {
-            *player = pl;
-        }
-        // if allowed player guid (if no then only online players allowed)
-        if (player_guid)
-        {
-            *player_guid = pl ? pl->GetObjectGuid() : ObjectGuid();
-        }
-
-        if (player_name)
-        {
-            *player_name = pl ? pl->GetName() : "";
-        }
-    }
-
-    // some from req. data must be provided (note: name is empty if player not exist)
-    if ((!player || !*player) && (!player_guid || !*player_guid) && (!player_name || player_name->empty()))
-    {
-        SendSysMessage(LANG_PLAYER_NOT_FOUND);
-        SetSentErrorMessage(true);
-        return false;
-    }
-
-    return true;
-}
-
-/**
- * @brief Extracts an account id from an argument or selected player fallback.
- *
- * @param args The argument text pointer to parse.
- * @param accountName Optional output for the resolved account name.
- * @param targetIfNullArg Optional output for the selected player when no argument is given.
- * @return uint32 The extracted account id, or 0 on failure.
- */
-uint32 ChatHandler::ExtractAccountId(char** args, std::string* accountName /*= NULL*/, Player** targetIfNullArg /*= NULL*/)
-{
-    uint32 account_id = 0;
-
-    ///- Get the account name from the command line
-    char* account_str = ExtractLiteralArg(args);
-
-    if (!account_str)
-    {
-        if (!targetIfNullArg)
-        {
-            return 0;
-        }
-
-        /// only target player different from self allowed (if targetPlayer!=NULL then not console)
-        Player* targetPlayer = getSelectedPlayer();
-        if (!targetPlayer)
-        {
-            return 0;
-        }
-
-        account_id = targetPlayer->GetSession()->GetAccountId();
-
-        if (accountName)
-        {
-            sAccountMgr.GetName(account_id, *accountName);
-        }
-
-        if (targetIfNullArg)
-        {
-            *targetIfNullArg = targetPlayer;
-        }
-
-        return account_id;
-    }
-
-    std::string account_name;
-
-    if (ExtractUInt32(&account_str, account_id))
-    {
-        if (!sAccountMgr.GetName(account_id, account_name))
-        {
-            PSendSysMessage(LANG_ACCOUNT_NOT_EXIST, account_str);
-            SetSentErrorMessage(true);
-            return 0;
-        }
-    }
-    else
-    {
-        account_name = account_str;
-        if (!Utf8ToUpperOnlyLatin(account_name))
-        {
-            PSendSysMessage(LANG_ACCOUNT_NOT_EXIST, account_name.c_str());
-            SetSentErrorMessage(true);
-            return 0;
-        }
-
-        account_id = sAccountMgr.GetId(account_name);
-        if (!account_id)
-        {
-            PSendSysMessage(LANG_ACCOUNT_NOT_EXIST, account_name.c_str());
-            SetSentErrorMessage(true);
-            return 0;
-        }
-    }
-
-    if (accountName)
-    {
-        *accountName = account_name;
-    }
-
-    if (targetIfNullArg)
-    {
-        *targetIfNullArg = NULL;
-    }
-
-    return account_id;
-}
-
-struct RaceMaskName
-{
-    char const* literal;
-    uint32 raceMask;
-};
-
-static RaceMaskName const raceMaskNames[] =
-{
-    // races
-    { "human", (1 << (RACE_HUMAN - 1))   },
-    { "orc", (1 << (RACE_ORC - 1))     },
-    { "dwarf", (1 << (RACE_DWARF - 1))   },
-    { "nightelf", (1 << (RACE_NIGHTELF - 1))},
-    { "undead", (1 << (RACE_UNDEAD - 1))  },
-    { "tauren", (1 << (RACE_TAUREN - 1))  },
-    { "gnome", (1 << (RACE_GNOME - 1))   },
-    { "troll", (1 << (RACE_TROLL - 1))   },
-    { "bloodelf", (1 << (RACE_BLOODELF - 1))},
-    { "draenei", (1 << (RACE_DRAENEI - 1)) },
-
-    // masks
-    { "alliance", RACEMASK_ALLIANCE },
-    { "horde",    RACEMASK_HORDE },
-    { "all", RACEMASK_ALL_PLAYABLE },
-
-    // terminator
-    { NULL, 0 }
-};
-
-/**
- * @brief Extracts a race mask from a numeric value or named preset.
- *
- * @param text The argument text pointer to parse.
- * @param raceMask Receives the resulting race mask.
- * @param maskName Optional output for the resolved preset name.
- * @return true if a race mask was extracted; otherwise false.
- */
-bool ChatHandler::ExtractRaceMask(char** text, uint32& raceMask, char const** maskName /*=NULL*/)
-{
-    if (ExtractUInt32(text, raceMask))
-    {
-        if (maskName)
-        {
-            *maskName = "custom mask";
-        }
-    }
-    else
-    {
-        for (RaceMaskName const* itr = raceMaskNames; itr->literal; ++itr)
-        {
-            if (ExtractLiteralArg(text, itr->literal))
-            {
-                raceMask = itr->raceMask;
-
-                if (maskName)
-                {
-                    *maskName = itr->literal;
-                }
-                break;
-            }
-        }
-
-        if (!raceMask)
-        {
-            return false;
-        }
-    }
-
-    return true;
 }
 
 /**
@@ -4326,78 +1837,6 @@ void ChatHandler::LogCommand(char const* fullcmd)
     {
         sLog.outCommand(GetAccountId(), "Command: %s [Account: %u from %s]",
                         fullcmd, GetAccountId(), GetAccountId() ? "RA-connection" : "Console");
-    }
-}
-
-void ChatHandler::BuildChatPacket(WorldPacket& data, ChatMsg msgtype, char const* message, Language language /*= LANG_UNIVERSAL*/, ChatTagFlags chatTag /*= CHAT_TAG_NONE*/,
-                                  ObjectGuid const& senderGuid /*= ObjectGuid()*/, char const* senderName /*= NULL*/,
-                                  ObjectGuid const& targetGuid /*= ObjectGuid()*/, char const* targetName /*= NULL*/,
-                                  char const* channelName /*= NULL*/)
-{
-    bool isGM = chatTag & CHAT_TAG_GM;
-
-    data.Initialize(isGM ? SMSG_GM_MESSAGECHAT : SMSG_MESSAGECHAT);
-    data << uint8(msgtype);
-    data << uint32(language);
-    data << ObjectGuid(senderGuid);
-    data << uint32(0);                                              // 2.1.0
-
-    switch (msgtype)
-    {
-        case CHAT_MSG_MONSTER_SAY:
-        case CHAT_MSG_MONSTER_PARTY:
-        case CHAT_MSG_MONSTER_YELL:
-        case CHAT_MSG_MONSTER_WHISPER:
-        case CHAT_MSG_RAID_BOSS_WHISPER:
-        case CHAT_MSG_RAID_BOSS_EMOTE:
-        case CHAT_MSG_MONSTER_EMOTE:
-            MANGOS_ASSERT(senderName);
-            data << uint32(strlen(senderName) + 1);
-            data << senderName;
-            data << ObjectGuid(targetGuid);                         // Unit Target
-            if (targetGuid && !targetGuid.IsPlayer() && !targetGuid.IsPet())
-            {
-                data << uint32(strlen(targetName) + 1);             // target name length
-                data << targetName;                                 // target name
-            }
-            MANGOS_ASSERT(message);
-            data << uint32(strlen(message) + 1);
-            data << message;
-            data << uint8(chatTag);
-            break;
-        case CHAT_MSG_BG_SYSTEM_NEUTRAL:
-        case CHAT_MSG_BG_SYSTEM_ALLIANCE:
-        case CHAT_MSG_BG_SYSTEM_HORDE:
-            data << ObjectGuid(targetGuid);                         // Unit Target
-            if (targetGuid && !targetGuid.IsPlayer())
-            {
-                MANGOS_ASSERT(targetName);
-                data << uint32(strlen(targetName) + 1);             // target name length
-                data << targetName;                                 // target name
-            }
-            MANGOS_ASSERT(message);
-            data << uint32(strlen(message) + 1);
-            data << message;
-            data << uint8(chatTag);
-            break;
-        default:
-            if (msgtype == CHAT_MSG_CHANNEL)
-            {
-                MANGOS_ASSERT(channelName);
-                data << channelName;
-            }
-            data << ObjectGuid(targetGuid);
-            MANGOS_ASSERT(message);
-            data << uint32(strlen(message) + 1);
-            data << message;
-            data << uint8(chatTag);
-            if (isGM)
-            {
-                MANGOS_ASSERT(senderName);
-                data << uint32(strlen(senderName) + 1);
-                data << senderName;
-            }
-            break;
     }
 }
